T442: CheckpointService as separate entity
