+++ src/main/scala/scorex/transaction/state/database/BlockStorageImpl.scala
@@ -7,7 +7,7 @@ import com.wavesplatform.state2.reader.StateReader
 import com.wavesplatform.state2.{BlockchainUpdaterImpl, MVStoreStateStorage, StateWriterImpl}
 import org.h2.mvstore.MVStore
 import scorex.transaction._
-import scorex.transaction.state.database.blockchain.{HistoryWriterImpl, MVStoreHistoryStorage}
+import scorex.transaction.state.database.blockchain.{CheckpointServiceImpl, HistoryWriterImpl, MVStoreCheckpointStorage, MVStoreHistoryStorage}
 
 class BlockStorageImpl(settings: BlockchainSettings) extends BlockStorage {
 
@@ -16,8 +16,8 @@ class BlockStorageImpl(settings: BlockchainSettings) extends BlockStorage {
   val database: MVStore = createMVStore(settings.file)
   val h = new HistoryWriterImpl(new MVStoreHistoryStorage(database))
   val s = new StateWriterImpl(new MVStoreStateStorage(database))
+  val c = new CheckpointServiceImpl(new MVStoreCheckpointStorage(database))
   val updater = new BlockchainUpdaterImpl(s, settings.functionalitySettings, h)
-
 
 
   override def history: History = h
@@ -26,7 +26,7 @@ class BlockStorageImpl(settings: BlockchainSettings) extends BlockStorage {
 
   override def blockchainUpdater: BlockchainUpdater = updater
 
-  override def checkpoints: CheckpointService = h
+  override def checkpoints: CheckpointService = c
 }
 
 object BlockStorageImpl {
