+++ src/main/scala/scorex/transaction/state/database/blockchain/HistoryStorage.scala
@@ -10,7 +10,6 @@ trait HistoryStorage {
   val blockIdByHeight: java.util.Map[Int, BlockId]
   val heightByBlockId: java.util.Map[BlockId, Int]
   val scoreByHeight: java.util.Map[Int, BigInt]
-  val checkpoint: java.util.Map[Int, Checkpoint]
 
   def commit(): Unit
 }
@@ -19,7 +18,6 @@ class MVStoreHistoryStorage(db: MVStore) extends HistoryStorage {
   val blockBodyByHeight: MVMap[Int, Array[Byte]] = db.openMap("blocks", new LogMVMapBuilder[Int, Array[Byte]])
   val blockIdByHeight: MVMap[Int, BlockId] = db.openMap("signatures", new LogMVMapBuilder[Int, BlockId])
   val heightByBlockId: MVMap[BlockId, Int] = db.openMap("signaturesReverse", new LogMVMapBuilder[BlockId, Int])
-  val checkpoint: MVMap[Int, Checkpoint] = db.openMap("checkpoint", new LogMVMapBuilder[Int, Checkpoint])
   val scoreByHeight: MVMap[Int, BigInt] = db.openMap("score", new LogMVMapBuilder[Int, BigInt])
 
   override def commit(): Unit = db.commit()
