+++ src/test/scala/scorex/network/BlockchainSynchronizerSpecification.scala
@@ -21,7 +21,7 @@ class BlockchainSynchronizerSpecification extends ActorTestingCommons {
 
   private def mockHistory(last: Int): History = {
     val history = mock[History]
-    (history.contains(_: BlockId)) expects * onCall { id: BlockId => id(0) <= last } anyNumberOfTimes()
+    (history.heightOf(_: BlockId)) expects * onCall { id: BlockId => if (id(0) <= last) Some(1) else None } anyNumberOfTimes()
     history
   }
 
@@ -304,7 +304,7 @@ class BlockchainSynchronizerSpecification extends ActorTestingCommons {
 
       def historyContaining(blockIds: Int*): History = {
         val history = mock[History]
-        (history.contains(_: BlockId)) expects * onCall { id: BlockId => blockIds.contains(id.head.toInt) } anyNumberOfTimes()
+        (history.heightOf(_: BlockId)) expects * onCall { id: BlockId => if (blockIds.contains(id.head.toInt)) Some(1) else None } anyNumberOfTimes()
         history
       }
